Reorganised StateController, sort code and tree command code
